Converter LibraryListView to TypeScript
